Splitted tools into validators and loaders
