+++ controller/cache_manager.go
@@ -7,11 +7,11 @@ import (
 	"time"
 
 	"api/bitcoin-api/models"
-	"api/bitcoin-api/tools"
+	"api/bitcoin-api/tools/loaders"
 )
 
 func writeCache(cache models.CachedPrice) error {
-	path, err := tools.GetEnvVariable("CACHE_PATH")
+	path, err := loaders.GetEnvVariable("CACHE_PATH")
 	if err != nil {
 		return fmt.Errorf("getting .env vavriable in writeCahce: %w", err)
 	}
@@ -29,7 +29,7 @@ func writeCache(cache models.CachedPrice) error {
 }
 
 func readCache() (*models.CachedPrice, error) {
-	path, err := tools.GetEnvVariable("CACHE_PATH")
+	path, err := loaders.GetEnvVariable("CACHE_PATH")
 	if err != nil {
 		return nil, fmt.Errorf("reading cache: %w", err)
 	}
