+++ controller/file_manager.go
@@ -6,11 +6,12 @@ import (
 	"strings"
 
 	"api/bitcoin-api/models"
-	"api/bitcoin-api/tools"
+	"api/bitcoin-api/tools/loaders"
+	"api/bitcoin-api/tools/validators"
 )
 
 func getEmails() ([]string, error) {
-	path, err := tools.GetEnvVariable("DATABASE_PATH")
+	path, err := loaders.GetEnvVariable("DATABASE_PATH")
 	if err != nil {
 		return nil, fmt.Errorf("extracting emails: %w", err)
 	}
@@ -48,7 +49,7 @@ func findEmail(email models.Email) (bool, error) {
 }
 
 func AddEmail(email models.Email) error {
-	path, err := tools.GetEnvVariable("DATABASE_PATH")
+	path, err := loaders.GetEnvVariable("DATABASE_PATH")
 	if err != nil {
 		return fmt.Errorf("getting .env variable: %w", err)
 	}
@@ -65,7 +66,7 @@ func AddEmail(email models.Email) error {
 		return fmt.Errorf("adding email %s: %w", email.Email, err)
 	}
 
-	if !tools.ValidateEmail(email.Email) || isPresent {
+	if !validators.ValidateEmail(email.Email) || isPresent {
 		return fmt.Errorf("provided email is invalid")
 	}
