+++ tools/loaders/env_loader.go
@@ -1,4 +1,4 @@
-package tools
+package loaders
 
 import (
 	"fmt"
